+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   History Worker web worker
+-   History indexedDB table
+-   Reply indexedDB table
+-   Reply table fallback for incognito users
 
 ### Changed